EENG-7475: Cleanup and comment in prep for review.
