+++ src/frontend/org/voltdb/iv2/MpRoSite.java
@@ -532,9 +532,4 @@ public class MpRoSite implements Runnable, SiteProcedureConnection
     public void applyBinaryLog(byte log[]) {
         throw new UnsupportedOperationException("RO MP Site doesn't do this, shouldn't be here");
     }
-
-    @Override
-    public void ensureProcLoaded(String procName) {
-        m_loadedProcedures.getProcByName(procName);
-    }
 }
